requirements.txt
litellm>=1.40
openai>=1.12
python-dotenv>=1.0
diskcache>=5.6
pyyaml>=6.0
jsonschema>=4.20
typer>=0.9
tqdm>=4.66
pytest>=7.4
numpy>=1.26
pandas>=2.1
sentence-transformers>=2.2
scikit-learn>=1.3
lightgbm>=4.1
xgboost>=2.0
optuna>=3.4
matplotlib>=3.8
seaborn>=0.13
scipy>=1.11
